feat: 텍스트분석모델 반영
